Docker: Add helpers for frequently used docker image/repository operations
The dockertest.images module defines several independent interfaces using
an abstract-base-class pattern.   They are extended through a few subclasses
to meet basic needs. There is an assumption that the full 64-character long
IDs are safer to use than the short, 12-character ones.  It is also assumed
that callers are in the best position to decide what if any 'tag' content
is used and how to mangle repo/image names.

Signed-off-by: Jiří Župka <jzupka@redhat.com>
Signed-off-by: Chris Evich <cevich@redhat.com>