images: Throw a real DockerCommandError
Some tests expect a DockerCommandError to be thrown
and will not accept a regular CmdError.  However,
this general behavior is wrong anyway because
callers should not be dependent on implementation-
specific exceptions.  Add a temporary band-aid
so DockerImagesCLI raises the expected exception.
Add a bunch of FIXME's to address the real problem
with some well-defined exception classes for common
conditions.

Signed-off-by: Chris Evich <cevich@redhat.com>