Docker Autotest: New control file example
Subtests and subsubtests which are to be disabled, if instantiated
here or in autotest **must** then report a result.  By design the
set of results does not include a "skip" or "disabled" code, since
this is the fundamental job of the control file.  With the exception
of subtests, being in the ``defaults.ini`` "disabled" list, these
changes prevent sub-subtests from loading.

Further, to help with the complex task of specifying which
sub/sub-subtests are available for running, a new control-file
is provided.  In addition to honoring the same ``--args`` syntax
as the default ``control`` file it adds the following features:

*  The ``--args`` argument is now treated as a order-sensitive
   space-separated list of values.  Any subtests or sub-subtest
   names listed alone or with commas (no spaces!) will be added
   to the master list of sub/sub-subtests to consider for testing.
   The sequence (including duplicates) is important, and preserved.

*  If the sub-option ``i=<thing>,<thing>,...`` appears in ``--args``,
   it will be used as the set of sub/sub-subtests to explicitly
   **include** from the master list above.  Unknown names are ignored,
   and an empty list means include everything.

*  If the sub-option ``x=<thing>,<thing>,...`` appears in ``--args``,
   it will be used as the set of sub/sub-subtests to explicitly
   **exclude** from the master list above.  Unknown names are ignored,
   and an empty list means include everything.  Items also appearing
   in include set, will be excluded.

For instances where a more static set of sub/sub-subtests, include
and exclude lists is needed, a completely optional control
configuration file is added:

*  A sample is provided in ``config_defaults/control.ini``, fully
   commented.

*  If the control file finds a ``config_custom/control.ini`` it
   will be loaded in preference to ``config_defaults/control.ini``.

*  The ``--args`` list of sub-subtests to consider will augment
   any sub/sub-subtests provided to the ``subtests`` option value.

*  The ``x=`` and ``i=`` sub-options will augment the ``include``
   and ``exclude`` option values.

Additionally, a new bugzilla-integration feature has been added
to help automatically skip tests with un-fixed bugs attached to them.
The mapping of sub/sub-subtest to bugzilla list is specified
in the ``control.ini`` file.  This feature is only enabled
if a bugzilla url option value is provided.  It is evaluated
last, after the normal include/exclude lists.

Signed-off-by: Chris Evich <cevich@redhat.com>