dockertest.subtests: Ignore empty/missing log args
Subtests.log_xn annoyingly complaining about multi-line
messages that don't have any arguments.  Since this is
a perfectly valid usage, the method needs to test the
contents of args, and simply ignore it when empty.  If
it's a problem, it will be obvious by the '%X' (tokens)
appearing in the output, no need to fail a test for this.

Signed-off-by: Chris Evich <cevich@redhat.com>