Support cascading __warning__ config. option
There are some global and test config options that
are highly recommended for site-specific customization.
For example, testing content that is difficult or
impractical to bundled in the open source releases.

Introduce a special __warning__ option that contains
a CSV list of all recommended for customization option
names (case insensitive).  This option's value does NOT
get overridden from one section to the next.  Instead,
each items are added to the list.

As part of final config processing, any config option
values which differ from the stock or default values,
are removed from the __warning__ list.  Later, during
subtest and sub-subtest initialization, a warning is issued
for any item remaining on this list.

Signed-off-by: Chris Evich <cevich@redhat.com>