Update tty option for all tests
The --tty run option ONLY works if a real tty/pty is attached
to the child-process stdin.  This is not the case for the vast
majority of tests.  Additionally Docker versions after 1.4 will
return an error if this option is used improperly.

Added a negativeusage test for the --tty option that expects
to receive the error code + message.

Removed BZ note from save_load test since it was resolved.

Signed-off-by: Chris Evich <cevich@redhat.com>