control: Additional test trees
The way inter-testing tasks are run isn't very flexable without
access to the dockertest package.  However, the control file should not
be dependent on the dockertest package anyway.  Add capability to execute
specific test modules before and after the main set of subtests.
In-between each subtest, execute a third set if 'intratests' tests.

* Modify test-name discovery to operate outside of subtests directory.

* Test steps are still produced up-front, however iteration over them
  is now a runtime activity.  This is much more effecient and will scale
  better as the entire testing state doesn't need to be updated every
  time.

* Test timeout is now derrived based on the global TIMEOUT set in the
  control file, divided up between all the subtests.  This guarantees
  that one test cannot monopolize the entire timeout.

* Added directories with examples for pretests, intratests, and
  posttests.

Signed-off-by: Chris Evich <cevich@redhat.com>