Atomic: Make build/install easier + modular
Previously everything was jammed into a single Dockerfile,
but a handful of scripts are also included.  Since the command
line for 'atomic install' is encoded in the Dockerfile,
it's more flexible to just call a script to finish the build.
This allows for more flexibility in base-image, install options,
and source locations.

Signed-off-by: Chris Evich <root@atomichost.local>