info subtest: fix verify_pool_name()
verify_pool_name() checks that the Docker Pool Name listed
by 'docker info' is present in the output of 'dmsetup ls'.
We cannot rely on the output order of the latter: on one
RHEL7 system it outputs the desired pool name first, on
another system last. So let's just test for the presence
of the desired pool name anywhere in the output.

Also, create a unit test for the new code. Invocation is
ugly but that should be fixable by someone more python-
proficient than I. Fixes welcome.

Unit test requires python-nose and python-unittest2, both
of which are available on RHEL7.

Signed-off-by: Ed Santiago <santiago@redhat.com>