Monkeypatch and add no-execute control argument
The Monkeypatch is used on sys.path to include the test directory in
module search path.  This is needed to support adding the
python-bugzilla module to the test-directory on systems where installing
elsewhere is impossible/impractical (i.e. Atomic).  The no-execute
argument is primarily intended for debugging control.ini and bugzilla
search options.

Signed-off-by: Chris Evich <cevich@redhat.com>