bz1097344 (docker-1.10 exit codes): prep work
Background: 'docker run' exit codes will be changing in 1.10,
to differentiate between container exit status and errors from
docker itself[1]. This commit prepares for that, by changing
the docker_expected_result config setting (PASS/FAIL)
to docker_expected_exit_status, an integer: 0 for success,
X for failure. As of this commit X is always 1, but at
some future point that will change. When it does,
the diagnostic message will help update the .ini setting
as necessary.

Additionally, the mustfail() helper now requires the
expected exit status as an argument. These statuses can
actually be hardcoded in the code itself, because
mustfail() tests (once) if the installed docker binary
returns unique exit codes; if it doesn't, it changes
the expectation to 1 (what old-docker always returns).

Also includes several unrelated fixes for minor problems
discovered during this process: removed an unused
config setting, fixed an incorrect test helper, renamed
a test that was colliding with an identically-named
test, and added (what I believe is) intended memoizing
to _client and _server in Version code.

 [1] https://github.com/docker/docker/pull/14012/files

Note that the exit codes here are UNTESTED, and will
require tweaking once we actually have docker-1.10
available. That is expected to be minor effort.

Signed-off-by: Ed Santiago <santiago@redhat.com>