Re-locate -D from client to server
The -D option causes docker client to issue mostly unhelpful
trace messages to stderr, interfering with tests that actually check
stderr.  Move it it to the default docker daemon command-line
configuration option.  That will cause any restarts of the docker
daemon to log additional details instead of spamming the console.
The daemon debug info. is also frequently requested by docker
engineers when examining possible bugs.

Also, the '-d' option for launching the daemon has been deprecated
since docker 1.8.  It's replacement is to pass the word 'daemon'
instead.  This was otherwise updated in init-scripts / systemd units,
but never reflected in the default docker autotest configuration.

Finally, removed the '--add-registry' option since it is going to
vary wildly from environment-to-environment.  The default in docker is
hard-coded to pull from docker.io which will work fine with the default
test image, so for the general case --add-registry is unnecessary.

Signed-off-by: Chris Evich <cevich@redhat.com>