commit subtest: fix race condition
Running docker with --detach, just to get a CID, exposes us to
a race condition if we check the container contents before the
command has completed.

Solution: use --name=<rand name> instead. That seems to be the
standard convention on other tests.

Also: clean up a diagnostic message: failif_ne() now displays
the expected and actual values, so caller doesn't have to

Signed-off-by: Ed Santiago <santiago@redhat.com>