docker 1.10 : new output for 'docker events'
docker 1.10 drastically changes the output format of the events command:

  - <timestamp> <sha> (...details...) <operation>
  + <timestmap> <container|image|network> <operation> <sha> (...details...)

This commit is a complete rewrite of the event-parsing code. One
goal is to support both formats. A more important goal is to
have robust parsing code with regression tests. The new code
defines strict expectations for what we should see, so we can
have confidence in what we parse (vs cherrypicking "search for
something that looks like a sha or fqin anywhere in the line").

Signed-off-by: Ed Santiago <santiago@redhat.com>