Update quickstart documentation and Prerequisites
The quickstart formerly recommended resetting back to the tag
of the current version.  However, this means any between-version
fixes are skipped.  Esp. between 0.8.4 and HEAD, in combination with
changes in docker's output will cause the version test to fail.
Because the latest master HEAD is more likely to be bug-free, it's a
much better starting-off point for anyone following the quickstart.

Added requrement for running as root, and changed the default path from
/usr/local/... to /var/lib/... because it's less problematic WRT selinux
default file contexts and certain tests (like run_volumes).

Updated the prerequisites section to include the 'Atomic' platforms
as well as make the runtime vs devel. software dependencies more distinct.
Docker Autotest is more than happy to run as-is on Atomic, but extra
bits are helpful for documentation and/or development work.

Signed-off-by: Chris Evich <cevich@redhat.com>