run_sigproxy: better diagnostics for timeout
Every so often, run_sigproxy/default fails with:

    Signal was not raised in container even though sig-proxy was enabled

Actual reason for this seems to be that the container just
isn't ready. Not sure why, or why only sometimes.

This commit tries to make that a little clearer by waiting
for a READY indicator. If we time out waiting for it, the
error message is a little more clear.

Signed-off-by: Ed Santiago <santiago@redhat.com>