DockerCmd : always run with verbose
When looking at a debug log of a failure, the first and
most important thing I want to know is: what docker commands
led to this? Too often it's not in the log, and I see this
in the source:

    DockerCmd(...., verbose=False)

...requiring an edit, and rerunning the test, then (when not
trivially reproducible) rerunning and rerunning some more.

Enough. Let's just always log docker commands. Since verbose
is the default, just remove most instances of verbose=False.
(And, while we're at it, also remove verbose=True since
it's a possibly-misleading NOP).

Also: fixed minor misspellings.

Signed-off-by: Ed Santiago <santiago@redhat.com>