get namespaces working
It actually turned out to be simple: just import fully-qualified
dockertest.<x> instead of <x>.

With this, the horrible sys.__path__ kludgery in the new __init__.py
is no longer necessary. Woot. And thanks, cevich, I cleaned up the
imports to remove 'import *'. I played with __all__ again but still
don't see any point to it other than adding duplication.

Code now working on docker-1.9 and -1.10. Unit tests working,
with addition of a PYTHONPATH=. in the unit test invoker.

Signed-off-by: Ed Santiago <santiago@redhat.com>