Extra run_pylint checking
It's been an unwritten rule that the python 'print'
statement shouldn't appear anywhere in module or test code.
Additionally, while used for debugging, the pdb module also
shouldn't appear in any checked-in code.  Update run_pylint
to with a function to check for both of these.

There is no easy way to add these checks to pylint/pep8.  In
the case of pylint and pdb, the workaround is to mark the pdb
module itself as deprecated instead.  This ensures all the possible
ways of importing it are checked.

In the case of print, the simplest solution is to just egrep
for it in every checked file.  If found, the file and line
references are printed.

Finally, I added a check for AUTOTEST_PATH or successful autotest
import, similar to the one that appears in run_unittests.sh.  I
also updated TMPFILENAME to use the mktemp tool instead of relying
on random numbers - which can clash.  Plus some string-formatting
fixes.

Signed-off-by: Chris Evich <cevich@redhat.com>