Treat TestNAError as PASS
When summarizing subtest results, treat TestNAError
(including DockerTestNAError) as a passing subsubtest.

FIXME: this could potentially mask real problems, such as
run_volumes skipping because of a wrong config setting or
one of the expecting-docker-1.10 tests skipping even though
the test system should have 1.10. One option is to define
a (global?) config setting such that TestNAError is treated
as fail. Another is to extend the test summary such that
there's a visual indication that some tests were skipped.
For now, we implement the la-la-la-not-listening option.

Signed-off-by: Ed Santiago <santiago@redhat.com>