run_sigproxy test: race condition fix
...at least, it's acting like a race condition (unpredictable
failures), even though on examination it looks like a pretty
clear bug that should fail every time.

The problem seems to be that some containers are run with --detach,
hence we don't see READY in stdout. Solution: use 'docker logs' in
addition to checking stdout. But why do the tests pass sometimes??

Signed-off-by: Ed Santiago <santiago@redhat.com>