refactor: make wait_for_ready() reusable
Other subtests (e.g. kill and possibly run_cidfile) are
experiencing spurious failures that look suspiciously like
the race condition that prompted wait_for_ready() in
run_sigproxy, i.e. we may be starting tests on containers
that aren't fully running yet.

This moves wait_for_ready() into a method of AsyncDockerCmd.
It's kind of complicated because we need to determine our
own container's ID, and that's not trivial or robust, but
for our purposes I think it's safe enough.

There is no corresponding method for DockerCmd, because it
only makes sense if invoked with --detach, and even in that
case there's no sane way to get the container ID.

In order to keep things simple, this is purely a refactor
with no extra functionality. If this code is acceptable, I'll
follow up with PRs for other failing tests.

Signed-off-by: Ed Santiago <santiago@redhat.com>