Pre-pull test images to save subsubtest timeout
It's easier for subtests to timeout if they're required to pull the test
image across a network of unknown capacity.  However, no test here is
exercizing network bandwidth, so any pull-related timeouts are failures
to be avoided.  The default control file imposes the subtest timeout
on pretests.  However pretests are not included in the calculation based
on the overall timeout.

Add a pre-test that pulls the default testing image, and optionally
additional images.  Also optionally (default: False) update
the config_custom/defaults.ini file (if it exists) with these images
to be ignored for cleaning by all subtests.

Finally, log information about all images that exist on the system into
the sysinfo results directory.  This makes it easy to find the initial
repository state of the system, including image ID's.

Signed-off-by: Chris Evich <cevich@redhat.com>

squash! Pre-pull test images to save subsubtest timeout

Correction: pretests are subject to global timeout