Loosen per-subtest timeout calculation
It's a safe assumption that most tests will finish before their
allocated timeout hits.  Otherwise, when the global timeout is too
short, testing is more likely to fail.  However, there isn't an
easy way to predict if/which tests will finish early.  Also, it's
impossible to re-allocate the unused timeouts to subsequent tests
(they are all created ahead of execution-time).  The best compromise is
to simply increase the per-test timeout by half as much.  If this is
wrong, then the global timeout will still ultimately apply, and kill off
the entire run.

Signed-off-by: Chris Evich <cevich@redhat.com>