Fix containers/images string-timeout
The config module tries it's best to interperate values, but sometimes
it's possible quotes sneak in or for whatever reason a string gets
handed back as a timeout when it should be an int or float.  Add another
layer of type-conversion / validation to make these heavily-used modules
more robust.  Also log a warning message to help make debugging further
problems easier.

Signed-off-by: Chris Evich <cevich@redhat.com>