Bump up some wait_for_ready timeouts
From 3-5s to 10s. Most of the time these subtests pass just fine,
but once in a while they fail while waiting for READY. This could
just be system load, I don't know what causes it. But extending
the timeout doesn't hurt, because wait_for_ready() triggers as
soon as the container issues READY -- it's not a sleep, we're
not wasting those 10 seconds. And with this extension we should
see fewer false alarms, which means less wasted developer time.

Signed-off-by: Ed Santiago <santiago@redhat.com>