Fix subtests requiring a Red Hat system
Add a simple way for tests to determine w/n they're running on a Red Hat
Enterprise Linux system (including Atomic).  In subtests where that is a
requirement, they should raise TestNAError.  Add unittests to verify the
new calls.  Update the two existing tests to make use of this feature.

Signed-off-by: Chris Evich <cevich@redhat.com>