Fix docker_test_images clobber defaults.ini (#646)
For whatever reason, ``dockertest.config.ConfigSection.merge_write()``
cannot properly merge with the ``DEFAULTS`` section.  Add a note/comment
to this effect, and update docker_test_images to perform the merge
with currently loaded defaults.  Also append a comment to the
updated file indicating the source of the modifications.  Did not investigate
"why" too deeply since fixing the bug is a high priority at the moment.

Signed-off-by: Chris Evich <cevich@redhat.com>