Replace run_ping with run_install
The run_ping test had a non-obvious dual-use of confirming installing
software was functional, and that basic networking is functional.
However, due to distribution base-image differences and setups, it's not
practical to squeeze both into a single container execution.

The replacement, run_install sub-subtest allows for this in a much
better way.  It provides two separate container executions, with the
later being from the image committed from the former.  If either the
installation fails, or the verification command fails, the test will
fail.  In all cases, the containers and committed image will be cleaned
up.