Prefetch centos, to fix failing build test (#659)
* Prefetch centos, to fix failing build test

build/https_file and build/rm_false tests are failing
because their Dockerfiles pull in centos, so img_count()
finds two new post-run images instead of one.

This patch solves it, but not ideally. It's duplication on top
of more duplication. A better solution might piggyback off
preserve_fqins but that too is fragile. Submitting as a
compromise to get tests passing.

Signed-off-by: Ed Santiago <santiago@redhat.com>