Strip out nonprintable characters from xml (#670)
Some test somewhere is generating an error message
with ANSI codes: ^[[91mpermission denied. This is
making Jenkins choke with invalid XML.

Strip off chars not in ``string.printable`` in production of the XML
data.  The autotest and client logs preserve the erroneous output
if inspecting it later is needed.  This is just about getting the
XML correct.

Signed-off-by: Ed Santiago <santiago@redhat.com>