Fix run_volumes_selinux depending on ls command
This test will fail if the --scontext option happens to not exist.
In any case, options to the 'ls' command could be different across
distros / versions or change unpredictably (WRT docker autotest).
Instead, add an API call to retrieve the context directly from the
extended attribute value.  This works on all filesystems and whether or
not SELinux is enabled.

Also fix a problem where, depending on the
underlying storage, testing files are not actually written to disk from
the host perspective.  Fixing this for all filesystems (including NFS)
requires writing some data to the file (not just touching it) and
calling sync from the host.

Signed-off-by: Chris Evich <cevich@redhat.com>