log_versions: handle missing rpms
A given system may have docker but not docker-latest (or,
rarely, vice-versa). Such situations lead to CmdError
failure in the log_versions pretest.

Solution: allow errors in 'rpm -q' command. If 'is not installed'
appears in command output, return empty string which will in turn
be written to the docker_rpm_xxx file.

Alternate solutions considered: returning the 'is not installed'
string itself, or dealing with it in write_sysinfo() by not
writing the docker_rpm_xxx file. IMO both of those make it
harder for client code.

Signed-off-by: Ed Santiago <santiago@redhat.com>