Fix masking exception in get_selinux_context()
The previous behavior of this function was to return None if the
file/directory did not exist.  This was a hold-over from when the
function was embedded in docker_cli/run_volumes_selinux.  However,
it's not ideal to continue this behavior in an API function because
it prevents use-cases that rely on an IOError being raised.

I've analyzed the run_volumes_selinux and there's no explicit references
to a ``None`` return value.  All calls to this function would
(appropriately) pass up an exception being raised from set_selinux_context().
Also, it was only a few commits ago this API function was added, so it's
safe to assume there are not likely any other callers anywhere.

Signed-off-by: Chris Evich <cevich@redhat.com>