Resolve set_selinux_context() FIXME
``pwd`` is misleading, it should be ``path`` - which also matches
``get_selinux_context()``.

Making ``pwd`` a keyword argument preserves compatibility with all
prior callers that used it as a keyword argument.  Since it was
the only positional argument, replacing it with a new first positional
keyword ``path`` argument also helps preserve behavior with old callers.

Having ``path`` default to ``None``, catches any new callers that use
keyword arguments, but fail to pass in ``pwd``.  In all cases, if ``pwd``
is used, issue a deprecation warning so callers don't use it
accidentally in new code.

Signed-off-by: Chris Evich <cevich@redhat.com>