info subtest: handle overlay2 driver
Up until now, the 'docker info' test has only had
to deal with devicemapper. Now there's overlay2,
and the docker info output for it is significantly
different. Much of the devicemapper validation is
not applicable to overlay2.

Split the storage-handler validation into individual
functions for each driver: currently devicemapper
and overlay2, each of which has different tunables.
As more back ends are added, we can add new functions
to check their settings.

Also: rewrite _build_table() so it handles sub-elements,
and add a unit test.

Signed-off-by: Ed Santiago <santiago@redhat.com>