Give Travis CI the boot
Too much effort has been spent trying to squeeze it's Ubuntu-based,
container and virtualenv environment into a RHEL-work-a-like.  Drop
Travis entirely, and instead rely on a TBD system for running the four
quick-check commands:

- SPHINXOPTS="-W" make
- ./run_checkdocs.py
- ./run_unittests.sh
- ./run_pylint.sh --FF

Replace the quick-code checking with jlebon/redhat-ci

Similar idea as travis, only it runs on RHEL, CentOS, and Fedora (or all
three at the same time).  This is also the same CI framework used by
most of the projectatomic repos.

Ref: https://github.com/jlebon/redhat-ci

Signed-off-by: Chris Evich <cevich@redhat.com>