Fix Broken BZ search by DEFAULTS section
When SafeConfigParser is initialized with a defaults dict,
it results in production of a ``[DEFAULTS]`` section when written.
For some reason this prevents properly connecting to bugzilla.
Possibly because the default keys are applied globally, to all sections.

Fix this by not passing a defaults dict, and instead hard-coding the
expected format (including sections) of this file.  This prevents
output of a ``[DEFAULTS]`` section and the BZ search happens normally.

Signed-off-by: Chris Evich <cevich@redhat.com>