Reimplement build test to better re-use methods.
Also, because the origin of the test image is largely unknown, it's
incorrect to assume it can always be ``docker pull``ed.  Yet it's also
not possible to properly disable docker's built-in "implied pull"
(for some non-existent local image).

Though the precise failure-path is unknown, the original (duplicated)
sub-subtest cleanup (somehow) allows the test images to be
be deleted (despite it being in the ``existing_images`` list).  Attempt
to fix this by:

* Move common methods into a new ``BuildBase`` class so they can be
reused by descendants.

* Make both ``build`` (the subtest), and ``BuildSubSubtest`` descendants
utilize common methods.

* Rewrite cleanup() to utilize the more common ``clean_all()`` methods
from the ``images`` and ``containers`` modules.  This will now respect
the ``preserve_*`` configuration options.

Signed-off-by: Chris Evich <cevich@redhat.com>