Fix build test image count offset
Previously when counting containers/images, the starting count
was from those that existed during subtest ``initialize()``.  However, if
a previous sub-subtest failed clean up a built, or additionally pulled
image(s), the count will be always be wrong (for all remaining sub-subtests).
Fix this by using the count of existing images found during ``initialize()``
of the sub-subtest.  This count will better reflect the actions of the
specific sub-subtest, instead of all those that came before.

Signed-off-by: Chris Evich <cevich@redhat.com>