deferred_deletion test: rewrite it (#721)
deferred_deletion test: rewrite it

Invalid assumption: we cannot rely on MountFlags=slave
being absent from docker systemd unit file. That's
only necessary for liverestore, not for deferred deletion.

Deferred deletion still works with MountFlags=slave,
we just can't access the container's filesystem from
inside the host. We have to nsenter docker's mount
namespace. And to do that we need a helper script.

Also: add one more prerequisite: check for the existence
of the sysctl knob file, and N/A if it's missing. This
covers the RHEL 7.2 case in which docker-storage-setup
optimistically and blindly enabled deferred_deletion.

Signed-off-by: Ed Santiago <santiago@redhat.com>