Fix build test examining stdout vs stderr
In the intermediate_exist_postprocess() and last_postprocess()
handlers, the build output is examned and searched to account for
created containers/images during the build.  After the build, any
leftover containers/images are compared to expectations given in
configuration.

However, these two methods were incorrectly searching stdout from the
build instead of stderr.  This (among other reasons) was causing the
bad_quiet sub-subtest to fail with a ``IndexError: list index out of
range``.  Though the underlying problem was the list of intermediate
containers was empty, this could easily also happen if a build
immediatly failed.

Fix both problems by examining stderr, and checking if the list was
empty before trying to reference inside it.

Signed-off-by: Chris Evich <cevich@redhat.com>