Include Fedora test image in-repo
Testing with a Fedora host requires using a special test image.
Formerly this was hosted/setup externally, but it's more convenient
to include it directly.  The default options for building it remain
commented out, and setting up the name in ``defaults.ini`` will
be documented in a later commit.

Also fix the order of test-image options in ``defaults.ini`` so
they follow the FQIN format sequence. i.e. host/user/name:tag

Signed-off-by: Chris Evich <cevich@redhat.com>