Simplify log_versions
Previously versions of the various docker RPMS was collected in two
places. This is redundant possibly ambiguous, and unnecessary.  The
"active" version of docker is also meaningless since subtests are free
to restart at will.  It would only represent the active version when the
pretest runs.  Better to leave making any declaration of "active"
version up to any subtests where it matters (none currently).

Signed-off-by: Chris Evich <cevich@redhat.com>