Workaround docker-latest build re: BZ 1403326
Docker 1.12+ regresses WRT this bug in that it adds additional
escape-codes to docker build output.  This is being (correctly)
caught by the ``OutputGood()`` check, however the bug seems
unlikely to ever be fixed.  Unfortunately, removing the check
isn't enough, because many sub-subtests also rely on regular-expression
matching against output.

For example, searching for ``^'Schazam!$'`` where the build output is:

```
Step 4/6 : RUN echo "Schazam!"
 ---> Running in cc59317feb51

[91m[0mSchazam!
 ---> 8a359f725df3
Removing intermediate container cc59317feb51
```

Would fail to match due to escape codes at the start of the important
line.  This commit adds a configurable option, defaulting to "off" which
enables checking for escape codes in ``OutputGood()``.  It also removes
the start-of-line match token from all the regex's, but leaves the
end-of-line token.  The assumption here is that escape codes will never
appear at the end of the line being matched (possibly not true).

Signed-off-by: Chris Evich <cevich@redhat.com>