wait_for_ready(): better handling of finished containers (#737)
In main timer loop, check container-running status (via self.done).
Terminate early if container is not running and we don't see the
expected READY string. This prevents a 10-minute wait on a
container that has exited with an error.

As long as we're in here, move the sleep() further down, to
the end of the loop. Not because it saves us 0.2 seconds but
because it makes the code flow more readable.

Signed-off-by: Ed Santiago <santiago@redhat.com>