images:remove_image_by_image_obj() - remove by ID, not name
Every so often the build subtest fails and leaves behind
an untagged intermediate image. This causes a cascade of
garbage_check failures:

    Found leftover images from prior test: [DockerImage(....)]

Root cause is that garbage_check _tries_ to remove these
images, but the underlying code was calling remove_by_fullname()
and fullname is undefined because TAG is <none>.

Solution: remove by image ID (hash), not name. This is reliable.

Signed-off-by: Ed Santiago <santiago@redhat.com>