Overlook known problems in docker builds (#761)
* Overlook known problems in docker builds

Certain subtests are known to fail under certain specific
docker builds. This is an inevitable part of life in a
fast-paced development environment.

Test failures are a distraction to humans: they waste time
and energy, often multiple days in a row because it's hard
to remember which failures we've seen before. This PR
introduces a lightweight mechanism for ignoring known
failures, by checking (subtest, docker-nvra) against
an easily-maintained list. When we find a match, we
change subtest failures to WARN and sub-subtest to PASS.

The key is an input file, known_failures.txt, inserted
into the config_custom subdirectory by ADEPT or other
test scaffolding. On any subtest or subsubtest failure
we read this file and cross-check the failing test
against the list.

Signed-off-by: Ed Santiago <santiago@redhat.com>